+++ src/main/java/techcourse/myblog/support/encryptor/EncryptHelper.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.support.util;
+package techcourse.myblog.support.encryptor;
 
 public interface EncryptHelper {
     String encrypt(String password);
