+++ src/test/java/techcourse/myblog/support/util/BCryptImplTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
+import techcourse.myblog.support.encryptor.EncryptHelper;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
